fix: drop graded problem test, as being outdated
Now that we're moving fully to the MFE, this older test will fail
once we remove the legacy views.

It did two things:
- visit the outline tab to click resume
- visit the courseware tab to click submit

Both legacy tabs are about to go away. If we want to reproduce
similar smoketest e2e checks, we may want to do that in the
cypress e2e suite instead.